+++ src/Services/PlanService.php
@@ -8,15 +8,15 @@ use Err0r\Larasub\Facades\PeriodService;
 final class PlanService
 {
     /**
-     * @param  \Err0r\Larasub\Models\Plan  $plan
+     * @param  \Err0r\Larasub\Models\PlanVersion  $planVersion
      * @param  Carbon  $startAt
      */
-    public function getPlanEndAt($plan, $startAt): ?Carbon
+    public function getPlanEndAt($planVersion, $startAt): ?Carbon
     {
         $endAt = null;
 
-        if ($plan->reset_period !== null && $plan->reset_period_type !== null) {
-            $endAt = $startAt->copy()->addDays(PeriodService::getDays($plan->reset_period, $plan->reset_period_type));
+        if ($planVersion->reset_period !== null && $planVersion->reset_period_type !== null) {
+            $endAt = $startAt->copy()->addDays(PeriodService::getDays($planVersion->reset_period, $planVersion->reset_period_type));
         }
 
         return $endAt;